refactor: Split NIP-01 message types into separate files.
