+++ src/nips/nip-0001-basic-protocol/events/outgoing-event-message-event.ts
@@ -5,7 +5,7 @@
  * @fileoverview Signifies a relay EVENT message on its way out to the client.
  */
 
-import { RelayEventMessage } from '../types/message-types';
+import { RelayEventMessage } from '../types/relay-event-message';
 import {
   ClientEvent,
   ClientEventOptions,
