+++ src/nips/nip-0001-basic-protocol/types/client-close-message.d.ts
@@ -1,0 +1,8 @@
+/**
+ * @license SPDX-License-Identifier: Apache-2.0
+ */
+/**
+ * @fileoverview A CLOSE message originating from a connected client.
+ */
+
+export type ClientCloseMessage = ['CLOSE', string];
