+++ src/nips/nip-0001-basic-protocol/types/client-req-messsage.d.ts
@@ -1,0 +1,10 @@
+/**
+ * @license SPDX-License-Identifier: Apache-2.0
+ */
+/**
+ * @fileoverview A REQ message originating from a connected client.
+ */
+
+import { Filter } from 'nostr-tools';
+
+export type ClientReqMessage = ['REQ', string, ...Filter[]];
