+++ src/nips/nip-0001-basic-protocol/types/generic-message.d.ts
@@ -1,0 +1,8 @@
+/**
+ * @license SPDX-License-Identifier: Apache-2.0
+ */
+/**
+ * @fileoverview A generic Nostr message.
+ */
+
+export type GenericMessage = [string, ...unknown[]];
