+++ src/nips/nip-0001-basic-protocol/types/relay-eose-message.d.ts
@@ -1,0 +1,8 @@
+/**
+ * @license SPDX-License-Identifier: Apache-2.0
+ */
+/**
+ * @fileoverview An EOSE message originating from the relay.
+ */
+
+export type RelayEOSEMessage = ['EOSE', string];
