+++ src/nips/nip-0001-basic-protocol/types/relay-event-message.d.ts
@@ -1,0 +1,10 @@
+/**
+ * @license SPDX-License-Identifier: Apache-2.0
+ */
+/**
+ * @fileoverview An EVENT message originating from the relay.
+ */
+
+import { Event as NostrEvent } from 'nostr-tools';
+
+export type RelayEventMessage = ['EVENT', string, NostrEvent];
